Re-organise spreadsheet generator files into main codebase
