+++ package.json
@@ -24,11 +24,11 @@
     "generatedata:allMaterials:withLinking": "node ./data/all-materials-data-generator.js --with-linking",
     "generatedata:allMaterialsMixed": "node ./data/all-materials-mixed-data-generator.js",
     "generatedata:allMaterialsMixed:withLinking": "node ./data/all-materials-mixed-data-generator.js --with-linking",
-    "generate:spreadsheet:input": "node ./data/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=reprocessorInput",
-    "generate:spreadsheet:output": "node ./data/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=reprocessorOutput",
-    "generate:spreadsheet:exporter": "node ./data/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=exporter",
-    "generate:spreadsheet:regOnlyExporter": "node ./data/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=regOnlyExporter",
-    "generate:spreadsheet:regOnlyReprocessor": "node ./data/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=regOnlyReprocessor",
+    "generate:spreadsheet:input": "node test/support/spreadsheet/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=reprocessorInput",
+    "generate:spreadsheet:output": "node test/support/spreadsheet/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=reprocessorOutput",
+    "generate:spreadsheet:exporter": "node test/support/spreadsheet/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=exporter",
+    "generate:spreadsheet:regOnlyExporter": "node test/support/spreadsheet/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=regOnlyExporter",
+    "generate:spreadsheet:regOnlyReprocessor": "node test/support/spreadsheet/summarylogs-spreadsheet-data-generator.js --wasteProcessingType=regOnlyReprocessor",
     "generate:spreadsheet:ors": "node ./data/ors-spreadsheet-generator.js",
     "test": "cucumber-js --tags \"not @smoketest and not @localonly\" ; npm run report",
     "test:smoketest": "cucumber-js --tags \"@smoketest\" ; npm run report",
