+++ test/support/spreadsheet/exporter.reg.only.js
